Fix narrowing conversion

+++ src/lib/storage/variable_string_dictionary/variable_string_dictionary_iterable.hpp
@@ -28,10 +28,10 @@ class VariableStringDictionarySegmentIterable
       using CompressedVectorIterator = decltype(vector.cbegin());
       using DictionaryIteratorType = decltype(_klotz->cbegin());
 
-      auto begin = Iterator<CompressedVectorIterator, DictionaryIteratorType>{_klotz->cbegin(), _klotz->size(),
-                                                                              vector.cbegin(), ChunkOffset{0u}};
+      auto begin = Iterator<CompressedVectorIterator, DictionaryIteratorType>{
+          _klotz->cbegin(), static_cast<uint32_t>(_klotz->size()), vector.cbegin(), ChunkOffset{0u}};
       auto end = Iterator<CompressedVectorIterator, DictionaryIteratorType>{
-          _klotz->cbegin(), _klotz->size(), vector.cend(), static_cast<ChunkOffset>(_segment.size())};
+          _klotz->cbegin(), static_cast<uint32_t>(_klotz->size()), vector.cend(), static_cast<ChunkOffset>(_segment.size())};
 
       functor(begin, end);
     });
@@ -49,10 +49,10 @@ class VariableStringDictionarySegmentIterable
 
           using PosListIteratorType = decltype(position_filter->cbegin());
           auto begin = PointAccessIterator<Decompressor, DictionaryIteratorType, PosListIteratorType>{
-              _klotz->cbegin(), _klotz->size(), vector.create_decompressor(), position_filter->cbegin(),
+              _klotz->cbegin(), static_cast<uint32_t>(_klotz->size()), vector.create_decompressor(), position_filter->cbegin(),
               position_filter->cbegin()};
           auto end = PointAccessIterator<Decompressor, DictionaryIteratorType, PosListIteratorType>{
-              _klotz->cbegin(), _klotz->size(), vector.create_decompressor(), position_filter->cbegin(),
+              _klotz->cbegin(), static_cast<uint32_t>(_klotz->size()), vector.create_decompressor(), position_filter->cbegin(),
               position_filter->cend()};
           functor(begin, end);
         });
